requirements.txt
# Core ML dependencies
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33

# Web framework
streamlit==1.28.1

# Image processing
pillow==10.0.1
numpy==1.24.3

# System monitoring
psutil==5.9.6

# Testing (optional for production)
pytest==7.4.3

# Additional dependencies for deployment
requests==2.31.0
urllib3==2.0.7